Refactored OpenAPI Info components.
